+++ magda-web-client/src/Components/DistributionRow.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
-import type { ParsedDistribution } from "../helpers/record";
+import { ParsedDistribution } from "../helpers/record";
 import { Link } from "react-router-dom";
 import { showTopNotification } from "../actions/topNotificationAction";
 import "./DistributionRow.css";
@@ -11,9 +11,12 @@ import newTabIcon from "../assets/external.svg";
 import { Medium } from "../UI/Responsive";
 import ga from "../analytics/googleAnalytics";
 import ReactTooltip from "react-tooltip";
+import AUbutton from "@gov.au/buttons";
+
 const formatIcons = {
     default: defaultFormatIcon
 };
+
 const dataFormatCategories = [
     "api",
     "archive",
@@ -219,7 +222,7 @@ class DistributionRow extends Component {
                 </div>
                 <div className="col-md-3 button-area">
                     {distribution.downloadURL ? (
-                        <button
+                        <AUbutton
                             className="download-button au-btn au-btn--secondary"
                             onClick={() => {
                                 if (!distribution.downloadURL) {
@@ -248,16 +251,9 @@ class DistributionRow extends Component {
                                 window.location = distribution.downloadURL;
                             }}
                         >
-                            <img src={downloadIcon} alt="download" />
-                            <a
-                                className="button-text"
-                                itemProp="contentUrl"
-                                href={distribution.downloadURL}
-                            >
-                                Download
-                            </a>
-                        </button>
-                    ) : null}{" "}
+                            Download <img src={downloadIcon} alt="download" />
+                        </AUbutton>
+                    ) : null}
                 </div>
             </div>
         );