+++ CHANGES.md
@@ -8,6 +8,7 @@
 -   Location and Date filters can be applied by clicking outside of the box
 -   Changed registry /records api `pageToken` as long type parameter to avoid runtime error
 -   Updated homepage tagline.
+-   Removed `a` link inside of dataset download button component.
 
 ## 0.0.42